market-guard/requirements.txt
Pillow
pandas
numpy
yfinance
fredapi
feedparser
vaderSentiment
python-dotenv
requests
resend

